@facundomedica 2022-09-05
```
facundo@Facundos-MacBook-Pro memdb % benchstat master-benchmark change-benchmark 
name                      old time/op  new time/op  delta
MemDBRangeScans1M-8        942µs ± 1%   192µs ± 2%  -79.58%  (p=0.000 n=19+17)
MemDBRangeScans10M-8       940µs ± 1%   216µs ± 6%  -76.99%  (p=0.000 n=18+17)
MemDBRandomReadsWrites-8  2.68µs ±16%  2.70µs ± 5%     ~     (p=0.464 n=20+17)
```