@alexanderbez 2022-09-05
@marbar3778 I think this PR is still worthwhile if it gives us some performance improvement and doesn't require too much effort.