@alexanderbez 2022-09-05
> > @marbar3778 I think this PR is still worthwhile if it gives us some performance improvement and doesn't require too much effort.
> 
> we don't use this code in the sdk right now. the change would need to land in cosmos/cosmos-db.

Yes we do. memDB is used by the CacheKVStore -- it powers the iterators.