@marbar3778 2022-09-05
> > > @marbar3778 I think this PR is still worthwhile if it gives us some performance improvement and doesn't require too much effort.
> > 
> > 
> > we don't use this code in the sdk right now. the change would need to land in cosmos/cosmos-db.
> 
> Yes we do. memDB is used by the CacheKVStore -- it powers the iterators.

 https://github.com/cosmos/cosmos-sdk/blob/ca86ec5e930b74305cafcb116376b2ea3e0473ec/store/cachekv/store.go#L10-L11

we use tm-db here. we don't use cosmos-sdk/db anywhere in the sdk currently. this is part of store/v2