@alexanderbez 2022-09-05
Might be worth it then to port to tm-db then.